Search results cleared after each search
